Fix typo's and small improvements
