+++ backend/fixtures/users.sql
@@ -1,2 +1,3 @@
 INSERT INTO users (id, username, password_hash)
+-- password is 'password'
 VALUES (1, 'user', '$argon2id$v=19$m=19456,t=2,p=1$frZGxFIhMHEsBJS4/VZr1A$zVIGEmiTFGy9jEy1Bphdq1ZO0lUngom8qu9PLsN6mZY');
